images retrieved from the repo in case of edit done

+++ app/src/main/java/com/example/task_rapidchidori_android/data/models/ImagesInfo.java
@@ -10,11 +10,11 @@ public class ImagesInfo {
     @PrimaryKey(autoGenerate = true)
     public int imageId;
 
-    public int taskId;
+    public long taskId;
 
     public String image;
 
-    public ImagesInfo(int taskId, @NonNull String image) {
+    public ImagesInfo(long taskId, @NonNull String image) {
         this.taskId = taskId;
         this.image = image;
     }
