+++ app/src/main/java/com/example/task_rapidchidori_android/ui/fragments/MyTasksFragment.java
@@ -310,18 +310,18 @@ public class MyTasksFragment extends Fragment implements View.OnClickListener, O
     @Override
     public void onItemClick(TaskInfo taskInfo) {
         Bundle bundle = new Bundle();
-        bundle.putInt(TASK_ID, taskInfo.taskID);
+        bundle.putLong(TASK_ID, taskInfo.taskID);
 
         Navigation.findNavController(requireActivity(), R.id.nav_host_fragment)
                 .navigate(R.id.action_myTasksFragment_to_addTaskFragment, bundle);
     }
 
     @Override
-    public void onTaskDelete(int taskId) {
+    public void onTaskDelete(long taskId) {
         showWarningDialog(taskId);
     }
 
-    private void showWarningDialog(int taskId) {
+    private void showWarningDialog(long taskId) {
         new AlertDialog.Builder(requireActivity())
                 .setTitle(R.string.delete_task_head)
                 .setMessage(R.string.delete_task_desc)
